Changed datatype of car ranges (electric range, conventional range, ...) from int to float Changed rounding of remaining[...]Range mathods from
(int) Math.floor(<e.g. fuel*maxRange>)

to

Math.floor(<e.g. fuel*maxRange> *1000.0f) /1000.0f so small rnages are not rounded to 0

Adapted unit tests and replaced expected results in example tests

+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/simulation/destinationChoice/CarRangeReachableZonesFilter.java
@@ -83,11 +83,11 @@ public class CarRangeReachableZonesFilter
 		private Set<Zone> filterByRangeOfCar(Set<Zone> zones, Person person, Zone currentZone, Zone nextFixedZone,
 				final float DIVERSION_FACTOR, final float RANGE_BUFFER_KM) {
 			Car car = person.whichCar();
-			Integer range = car.effectiveRange();
+			float range = car.effectiveRange();
 
 			Zone homeZone = person.household().homeZone();
 
-			if (range == Integer.MAX_VALUE) {
+			if (range == Float.MAX_VALUE) {
 				return zones;
 			}
